Refactorización de validaciones y mejora en eliminación de estudiantes: - Se agregó la función genérica validateUpdatesInputs para validar cédula y carnet antes de actualizar. - Se implementó la validación en actualizarEstudiante para evitar duplicados de cédula y carnet. - Se mejoró el manejo de errores con CustomError al detectar conflictos en la actualización. - Se añadió lógica para asegurar que, al eliminar un estudiante, también se elimine correctamente la persona, dirección y usuario asociado (corrigiendo el flujo de eliminación). - Se mejoraron los logs de depuración para el seguimiento de los datos procesados y los errores durante la eliminación. - Se corrigieron asociaciones y onDelete en modelos relevantes para asegurar eliminación en cascada correcta.
