+++ server.js
@@ -9,7 +9,7 @@ const sequelize = require('./src/config/database');
 const casosRoutes = require('./src/routes/casos');
 const usuarioRoutes = require('./src/routes/usuario');
 const estudiantesRoutes = require('./src/routes/estudiantes');
-const asignacionesRoutes = require('./src/routes/asignaciones');
+const authRoutes = require('./src/routes/auth');
 
 const app = express();
 const router = express.Router();
@@ -40,8 +40,8 @@ app.use(session({
 // Rutas
 app.use('/api/v1/usuarios', usuarioRoutes);
 app.use('/api/v1/casos', casosRoutes);
-app.use('/api/v1/asignaciones', asignacionesRoutes);
 app.use('/api/v1/estudiantes', estudiantesRoutes);
+app.use('/api/v1/auth', authRoutes);
 
 app.get('/api/v1/version', (req, res) => {
     res.send('v0.0.1');
