+++ src/controllers/validations/validations.js
@@ -1,6 +1,7 @@
 // Función de validación
 
-const { MESSAGE_ERROR } = require("../../constants/constants");
+const { MESSAGE_ERROR, HttpStatus } = require("../../constants/constants");
+const { CustomError } = require("../../handlers/responseHandler");
 
 const validateLoginInput = (email, password) => {
     const errors = [];
@@ -8,8 +9,26 @@ const validateLoginInput = (email, password) => {
     if (!password) errors.push(MESSAGE_ERROR.PASSWORD_IS_REQUIRED);
     return errors;
 };
+const validateUpdatesInputs = async ({ currentValue, newValue, model, field, message }) => {
+    // Si el valor ha cambiado, realiza la validación
+    if (currentValue !== newValue) {
+        const whereClause = {};
+        whereClause[field] = newValue; // Asigna dinámicamente el campo a validar
+
+        // Verifica si el nuevo valor ya existe en la base de datos
+        const existing = await model.findOne({
+            where: whereClause
+        });
+
+        if (existing) {
+            throw new CustomError(HttpStatus.BAD_REQUEST, message);
+        }
+    }
+}
 
 
 
 
-module.exports = { validateLoginInput };
+
+
+module.exports = { validateLoginInput, validateUpdatesInputs };
