+++ src/models/estudiante.js
@@ -29,7 +29,7 @@ module.exports = (sequelize, DataTypes) => {
     // Definir asociaciones
     Estudiante.associate = models => {
         Estudiante.belongsTo(models.Persona, { // Un estudiante pertenece a una persona
-            foreignKey: TABLE_FIELDS.UID_ESTUDIANTE,
+            foreignKey: TABLE_FIELDS.UID_ESTUDIANTE,  //tenia -> TABLE_FIELDS.UID_ESTUDIANTE
             targetKey: TABLE_FIELDS.UID_PERSONA,
             onDelete: 'CASCADE',
         });
